+++ CHANGELOG.md
@@ -15,6 +15,9 @@
   `MollieEx.Balances.primary/2`.
 - Add `MollieEx.Balances.list_transactions/3`.
 - Add `MollieEx.Balances.get_report/5`.
+- Add `MollieEx.BalanceTransfers.create/3`,
+  `MollieEx.BalanceTransfers.list/2`, and
+  `MollieEx.BalanceTransfers.get/3`.
 
 ## 0.4.0 - 2026-06-14